pyproject.toml
[tool.poetry]
name = "CommonRoad-Carla-Interface"
version = "0.1.0"
description = "Interface between CommonRoad and the traffic simulator CARLA."
authors = ["Cyber-Physical Systems Group, Technical University of Munich <commonroad@lists.lrz.de>"]
license = "TODO"
readme = "README.md"
homepage = "https://commonroad.in.tum.de"
keywords= ["autonomous", "automated", "vehicles", "driving", "motion", "planning", "simulation"]
classifiers = [
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Operating System :: POSIX :: Linux",
]
packages = [{include = "crcarla"}]

#[tool.poetry.urls]
#Documentation = "https://commonroad.in.tum.de/static/docs/commonroad-io/index.html"
#Forum = "https://commonroad.in.tum.de/forum/c/commonroad-io"
#Source = "https://gitlab.lrz.de/tum-cps/commonroad_io"


[tool.poetry.dependencies]
python = ">=3.9,<3.11"
pygame = ">=2.1.2"
commonroad-io = "2023.4"
carla = "0.9.15"
omegaconf = ">=2.1.2"
commonroad-drivability-checker = "~=2023.1"
commonroad-vehicle-models = "~=3.0.2"

psutil = "^5.9.4"
commonroad-scenario-designer = { git = "git@gitlab.lrz.de:cps/commonroad-scenario-designer.git", branch = "cr2odr" }
commonroad-reactive-planner = { git = "git@gitlab.lrz.de:cps/reactive-planner.git", branch = "feature/carla" }
commonroad-prediction = ">=0.1.0"
numpy = [
    { version = ">=1.22, <1.26", python = "3.8" },
    { version = "^1.26.1", python = ">=3.9,<3.13" }
]
scipy = [
    { version = ">=1.10.0,<1.11", python = "3.8" },
    { version = "^1.11.3", python = ">=3.9,<3.13" }
]

#[tool.poetry.group.tests]
#optional = true
#
#[tool.poetry.group.tests.dependencies]

[tool.poetry.group.docs]
optional = true

[tool.poetry.group.docs.dependencies]
mkdocstrings-python = "^1.8.0"
mkdocs-material = "^9.5.10"


[tool.poetry.group.dev]
optional = true

[tool.poetry.group.dev.dependencies]
black = "24.1.1"
isort = "5.13.1"
flake8 = "6.1.0"
pre-commit = "3.5.0"

[tool.black]
extend-exclude = "(gen/)"
line-length = 120

[tool.isort]
profile = "black"
extend_skip = ["gen"]

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
